tests: address review comments on PR #663
- Drop changeset (test-only, no release needed)
- Merge client edge-case tests into test_client.py, remove trivially-true
  test and unused imports
- Keep only the 5 handler_status_from_exit_command tests on the server
  side, move them into test_durable_runtime.py
- Move _get_persistence/_get_idle_release to server_test_fixtures.py
- Delete test_client_edge_cases.py and test_persistence_runtime_units.py
