--- .changeset/expand-client-and-persistence-runtime-test-coverage.md
@@ -1,6 +1,0 @@
----
-"llama-agents-client": patch
-"llama-agents-server": patch
----
-
-Expand test coverage for WorkflowClient error paths and TickPersistenceDecorator branches
